requirements.txt
numpy==1.18.5
matplotlib==3.3.0
Pillow==7.2.0
idx2numpy==1.2.2
opencv-python==4.4.0.42
pandas==1.1.1
notebook==6.1.1